Move functions in web.ex to helper modules
