+++ lib/enum_helper.ex
@@ -1,0 +1,7 @@
+defmodule Constable.EnumHelper do
+  def pluck(enumerable, property) do
+    Enum.map(enumerable, fn(object) ->
+      Map.get(object, property)
+    end)
+  end
+end
